requirements.txt
pytorch==1.7.0
torchvision==0.10.0
prettytable=3.9.0